@@ -1,52 +1,34 @@
 {
   "private": true,
-  "packageManager": "pnpm@8.2.0",
-  "engines": {
-    "node": ">=18",
-    "pnpm": ">=8"
-  },
   "scripts": {
-    "dev": "next dev --turbo",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "lint-staged": "lint-staged",
     "prettier": "prettier --write --ignore-unknown .",
     "prettier:check": "prettier --check --ignore-unknown .",
-    "test": "pnpm lint && pnpm prettier:check"
-  },
-  "git": {
-    "pre-commit": "lint-staged"
-  },
-  "lint-staged": {
-    "*": "prettier --write --ignore-unknown"
+    "test": "pnpm prettier:check"
   },
   "dependencies": {
-    "@headlessui/react": "^1.7.19",
-    "@heroicons/react": "^2.1.3",
-    "clsx": "^2.1.0",
-    "geist": "^1.3.0",
-    "next": "14.2.2",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "@headlessui/react": "^2.2.0",
+    "@heroicons/react": "^2.2.0",
+    "clsx": "^2.1.1",
+    "geist": "^1.3.1",
+    "next": "15.0.4",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "sonner": "^1.7.0"
   },
   "devDependencies": {
     "@tailwindcss/container-queries": "^0.1.1",
-    "@tailwindcss/typography": "^0.5.12",
-    "@types/node": "20.12.7",
-    "@types/react": "18.2.79",
-    "@types/react-dom": "18.2.25",
-    "@vercel/git-hooks": "^1.0.0",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^8.57.0",
-    "eslint-config-next": "^14.2.2",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-unicorn": "^52.0.0",
-    "lint-staged": "^15.2.2",
-    "postcss": "^8.4.38",
-    "prettier": "3.2.5",
-    "prettier-plugin-tailwindcss": "^0.5.14",
-    "tailwindcss": "^3.4.3",
-    "typescript": "5.4.5"
+    "@tailwindcss/typography": "^0.5.15",
+    "@types/node": "22.10.1",
+    "@types/react": "19.0.0",
+    "@types/react-dom": "19.0.1",
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.4.49",
+    "prettier": "3.4.2",
+    "prettier-plugin-tailwindcss": "^0.6.9",
+    "tailwindcss": "^3.4.16",
+    "typescript": "5.7.2"
   }
 }
