@@ -6,7 +6,7 @@ import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { useEffect, useState } from 'react';
 
-const FooterMenuItem = ({ item }: { item: Menu }) => {
+export function FooterMenuItem({ item }: { item: Menu }) {
   const pathname = usePathname();
   const [active, setActive] = useState(pathname === item.path);
 
@@ -29,7 +29,7 @@ const FooterMenuItem = ({ item }: { item: Menu }) => {
       </Link>
     </li>
   );
-};
+}
 
 export default function FooterMenu({ menu }: { menu: Menu[] }) {
   if (!menu.length) return null;
